refactor: replace the use of z-navigation-tabs subcomponents with native elements (for a11y)
