+++ src/components/icons/z-icon/readme.md
@@ -59,9 +59,9 @@
  - [z-myz-card-icon](../../../snowflakes/myz/card/z-myz-card-icon)
  - [z-myz-card-info](../../../snowflakes/myz/card/z-myz-card-info)
  - [z-myz-list-item](../../../snowflakes/myz/list/z-myz-list-item)
- - [z-navigation-tab](../../navigation/tabs/z-navigation-tab)
- - [z-navigation-tab-link](../../navigation/tabs/z-navigation-tab-link)
- - [z-navigation-tabs](../../navigation/tabs/z-navigation-tabs)
+ - [z-navigation-tab](../../../deprecated/z-navigation-tab)
+ - [z-navigation-tab-link](../../../deprecated/z-navigation-tab-link)
+ - [z-navigation-tabs](../../navigation/z-navigation-tabs)
  - [z-notification](../../notification/z-notification)
  - [z-pagination](../../z-pagination)
  - [z-panel-elem](../../panel/z-panel-elem)
