--- src/components/navigation/tabs/z-navigation-tabs/readme.md
@@ -1,37 +1,0 @@
-# z-navigation-tabs
-
-<!-- Auto Generated Below -->
-
-
-## Properties
-
-| Property      | Attribute     | Description                               | Type                                                                         | Default                                |
-| ------------- | ------------- | ----------------------------------------- | ---------------------------------------------------------------------------- | -------------------------------------- |
-| `ariaLabel`   | `aria-label`  | Set aria-label attribute in tablist role. | `string`                                                                     | `undefined`                            |
-| `orientation` | `orientation` | Navigation tabs orientation.              | `NavigationTabsOrientation.HORIZONTAL \| NavigationTabsOrientation.VERTICAL` | `NavigationTabsOrientation.HORIZONTAL` |
-| `size`        | `size`        | Navigation tabs size.                     | `NavigationTabsSize.BIG \| NavigationTabsSize.SMALL`                         | `NavigationTabsSize.BIG`               |
-
-
-## Slots
-
-| Slot | Description                                                                          |
-| ---- | ------------------------------------------------------------------------------------ |
-|      | Main slot. Use `z-navigation-tab` or `z-navigation-tab-link` components as children. |
-
-
-## Dependencies
-
-### Depends on
-
-- [z-icon](../../../icons/z-icon)
-
-### Graph
-```mermaid
-graph TD;
-  z-navigation-tabs --> z-icon
-  style z-navigation-tabs fill:#f9f,stroke:#333,stroke-width:4px
-```
-
-----------------------------------------------
-
-*Built with [StencilJS](https://stenciljs.com/)*
