--- src/components/navigation/tabs/z-navigation-tabs/styles.css
@@ -1,102 +1,0 @@
-:host {
-  --z-navigation-tabs-nav-buttons-bg: var(--color-white);
-  --z-navigation-tabs-nav-buttons-fg: var(--color-primary01);
-
-  position: relative;
-  z-index: 0;
-  display: flex;
-  overflow: hidden;
-  flex-direction: row;
-  font-family: var(--font-family-sans);
-  font-weight: var(--font-rg);
-}
-
-:host,
-:host *,
-::slotted(*) {
-  box-sizing: border-box;
-}
-
-::-webkit-scrollbar {
-  display: none;
-}
-
-.navigation-button {
-  position: absolute;
-  z-index: 1;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 0;
-  border: none;
-  margin: 0;
-  background: var(--z-navigation-tabs-nav-buttons-bg);
-  border-radius: var(--border-no-radius);
-  box-shadow: 0 0 4px 1px var(--shadow-color-base);
-  cursor: pointer;
-  fill: var(--z-navigation-tabs-nav-buttons-fg);
-  outline: none;
-}
-
-.navigation-button:disabled {
-  display: none;
-}
-
-nav {
-  display: flex;
-  overflow: auto;
-  align-items: center;
-  justify-content: flex-start;
-  scroll-behavior: smooth;
-
-  /* hide scrollbar in Firefox */
-  scrollbar-width: none;
-}
-
-:host([orientation="horizontal"]) nav {
-  width: 100%;
-}
-
-:host([orientation="horizontal"]) .navigation-button {
-  top: 0;
-  width: calc(var(--space-unit) * 4);
-  height: 100%;
-}
-
-:host([orientation="horizontal"]) .navigation-button:first-child {
-  left: 0;
-}
-
-:host([orientation="horizontal"]) .navigation-button:last-child {
-  right: 0;
-}
-
-/* Orientation vertical */
-:host([orientation="vertical"]) {
-  width: fit-content;
-  flex-direction: column;
-}
-
-:host([orientation="vertical"]) nav {
-  height: 100%;
-  flex-direction: column;
-  align-items: stretch;
-}
-
-:host([orientation="vertical"]) .navigation-button {
-  left: 0;
-  width: 100%;
-  height: calc(var(--space-unit) * 4);
-}
-
-:host([orientation="vertical"]) .navigation-button:first-child {
-  top: 0;
-}
-
-:host([orientation="vertical"]) .navigation-button:last-child {
-  bottom: 0;
-}
-
-:host([size="small"][orientation="vertical"]) .navigation-button {
-  height: calc(var(--space-unit) * 4);
-}
