+++ src/components/navigation/z-app-header/index.tsx
@@ -394,6 +394,8 @@ export class ZAppHeader {
             aria-label="Chiudi menu"
             onClick={this.closeDrawer}
             slot="canvasContent"
+            aria-hidden={!this.drawerOpen}
+            disabled={!this.drawerOpen}
           >
             <z-icon name="close"></z-icon>
           </button>
