+++ src/components/navigation/z-navigation-tabs/styles.css
@@ -1,0 +1,212 @@
+:host {
+  --z-navigation-tabs-nav-buttons-bg: ;
+  --z-navigation-tabs-nav-buttons-fg: ;
+
+  position: relative;
+  z-index: 0;
+  display: flex;
+  overflow: hidden;
+  flex-direction: row;
+  font-family: var(--font-family-sans);
+  font-weight: var(--font-rg);
+}
+
+:host,
+:host *,
+::slotted(*) {
+  box-sizing: border-box;
+}
+
+::-webkit-scrollbar {
+  display: none;
+}
+
+.navigation-button {
+  position: absolute;
+  z-index: 1;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 0;
+  border: none;
+  margin: 0;
+  background-color: var(--z-navigation-tabs-nav-buttons-bg, var(--color-surface01));
+  border-radius: var(--border-no-radius);
+  box-shadow: 0 0 4px 1px var(--shadow-color-base);
+  cursor: pointer;
+  fill: var(--z-navigation-tabs-nav-buttons-fg, var(--color-primary01));
+  outline: none;
+}
+
+.navigation-button:disabled {
+  display: none;
+}
+
+nav {
+  z-index: 0;
+  display: flex;
+  overflow: auto;
+  align-items: center;
+  justify-content: flex-start;
+  scroll-behavior: smooth;
+
+  /* hide scrollbar in Firefox */
+  scrollbar-width: none;
+}
+
+:host([orientation="horizontal"]) nav {
+  width: 100%;
+}
+
+:host([orientation="horizontal"]) .navigation-button {
+  top: 0;
+  width: calc(var(--space-unit) * 4);
+  height: 100%;
+}
+
+:host([orientation="horizontal"]) .navigation-button:first-child {
+  left: 0;
+}
+
+:host([orientation="horizontal"]) .navigation-button:last-child {
+  right: 0;
+}
+
+/* Orientation vertical */
+:host([orientation="vertical"]) {
+  width: fit-content;
+  flex-direction: column;
+}
+
+:host([orientation="vertical"]) nav {
+  height: 100%;
+  flex-direction: column;
+  align-items: stretch;
+}
+
+:host([orientation="vertical"]) .navigation-button {
+  left: 0;
+  width: 100%;
+  height: calc(var(--space-unit) * 4);
+}
+
+:host([orientation="vertical"]) .navigation-button:first-child {
+  top: 0;
+}
+
+:host([orientation="vertical"]) .navigation-button:last-child {
+  bottom: 0;
+}
+
+:host([size="small"][orientation="vertical"]) .navigation-button {
+  height: calc(var(--space-unit) * 4);
+}
+
+/* Style for the tabs */
+
+nav > ::slotted(*) {
+  position: relative;
+  z-index: 0;
+  display: inline-flex;
+  width: auto;
+  align-items: center;
+  justify-content: center;
+  padding: calc(var(--space-unit) * 2) calc(var(--space-unit) * 5);
+  border: none;
+  margin: 0;
+  background-color: unset;
+  border-radius: var(--border-no-radius);
+  color: var(--color-primary01);
+  cursor: pointer;
+  fill: currentcolor;
+  font-family: var(--font-family-sans);
+  font-size: inherit;
+  letter-spacing: inherit;
+  line-height: inherit;
+  outline: none;
+  text-align: center;
+  white-space: nowrap;
+}
+
+nav > ::slotted(a) {
+  text-decoration: none;
+}
+
+nav > ::slotted(*:focus:focus-visible) {
+  z-index: 1;
+  box-shadow: inset 0 0 4px 3px var(--color-highlight);
+}
+
+nav > ::slotted(:not([disabled]):hover) {
+  background-color: var(--color-surface02);
+}
+
+nav > ::slotted(button[disabled]) {
+  color: var(--color-disabled03);
+  cursor: not-allowed;
+  fill: currentcolor;
+  pointer-events: all;
+}
+
+nav > ::slotted(:not([disabled]):hover),
+nav > ::slotted([aria-selected="true"]) {
+  color: var(--color-hover-secondary);
+  fill: currentcolor;
+}
+
+nav > ::slotted(:not([disabled]):hover)::after,
+nav > ::slotted([aria-selected="true"])::after {
+  position: absolute;
+  background-color: var(--color-hover-secondary);
+  content: "";
+}
+
+:host([orientation="horizontal"]) nav > ::slotted(:not([disabled]):hover)::after,
+:host([orientation="horizontal"]) nav > ::slotted([aria-selected="true"])::after {
+  bottom: 0;
+  left: 0;
+  width: 100%;
+  height: var(--border-size-large);
+}
+
+nav > ::slotted(*) z-icon {
+  --z-icon-width: calc(var(--space-unit) * 2);
+  --z-icon-height: calc(var(--space-unit) * 2);
+
+  display: flex;
+  margin: 0;
+}
+
+:host([orientation="horizontal"]) nav > ::slotted(*) z-icon {
+  margin-right: var(--space-unit);
+}
+
+/* `small` size (only available for horizontal orientation) */
+
+:host([size="small"][orientation="horizontal"]) nav > ::slotted(*) {
+  padding: var(--space-unit) calc(var(--space-unit) * 2);
+}
+
+:host([size="small"][orientation="horizontal"]) nav > ::slotted(:not([disabled]):hover)::after,
+:host([size="small"][orientation="horizontal"]) nav > ::slotted([aria-selected="true"])::after {
+  height: var(--border-size-medium);
+}
+
+:host([size="small"]:not([orientation="vertical"])) nav > ::slotted(*) z-icon {
+  --z-icon-width: 14px;
+  --z-icon-height: 14px;
+}
+
+/* Orientation Vertical */
+
+:host([orientation="vertical"]) nav > ::slotted(*) {
+  padding: calc(var(--space-unit) * 3) calc(var(--space-unit) * 2);
+}
+
+:host([orientation="vertical"]) nav > ::slotted(:not([disabled]):hover)::after,
+:host([orientation="vertical"]) nav > ::slotted([aria-selected="true"])::after {
+  top: 0;
+  right: 0;
+  width: var(--border-size-large);
+  height: 100%;
+}
