+++ src/deprecated/z-navigation-tab-link/index.stories.mdx
@@ -1,13 +1,15 @@
 import {html} from "lit-html";
 import {Meta, Story, Canvas, ArgsTable} from "@storybook/addon-docs";
-import {NavigationTabsSize, NavigationTabsOrientation} from "../../../../beans/index";
-import {ICONS} from "../../../icons/icons";
+import {NavigationTabsSize, NavigationTabsOrientation} from "../../beans";
+import {ICONS} from "../../components/icons/icons";
 import "./index";
+
+⚠️ This component is deprecated. Replace it with a native `<a>` element.
 
 # ZNavigationTabLink
 
 <Meta
-  title="ZNavigationTabs/ZNavigationTabLink"
+  title="Deprecated/ZNavigationTabLink"
   component="z-navigation-tab-link"
   argTypes={{
     selected: {
