+++ src/deprecated/z-navigation-tab/index.stories.mdx
@@ -1,13 +1,15 @@
 import {html} from "lit-html";
 import {Meta, Story, Canvas, ArgsTable} from "@storybook/addon-docs";
-import {NavigationTabsSize, NavigationTabsOrientation} from "../../../../beans/index";
-import {ICONS} from "../../../icons/icons";
+import {NavigationTabsSize, NavigationTabsOrientation} from "../../beans";
+import {ICONS} from "../../components/icons/icons";
 import "./index";
+
+⚠️ This component is deprecated. Replace it with a native `<button>` element and use the `aria-controls` attribute to link it to the panel it controls.
 
 # ZNavigationTab
 
 <Meta
-  title="ZNavigationTabs/ZNavigationTab"
+  title="Deprecated/ZNavigationTab"
   component="z-navigation-tab"
   argTypes={{
     selected: {
