+++ src/deprecated/z-navigation-tab/readme.md
@@ -29,15 +29,13 @@
 | -------------- | --------------- | ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ---------------------------------------------------------------------------- | -------------------------------------- |
 | `ariaControls` | `aria-controls` | `aria-controls` attribute of the tab. Identifies the element (with `role=tabpanel`) whose contents or presence are controlled by this tab. The value must be the `id` of the element it controls.          | `string`                                                                     | `undefined`                            |
 | `disabled`     | `disabled`      | Whether the tab is disabled.                                                                                                                                                                               | `boolean`                                                                    | `false`                                |
-| `href`         | `href`          | Url to set to the anchor element.                                                                                                                                                                          | `string`                                                                     | `undefined`                            |
-| `htmlTitle`    | `html-title`    | Html title attribute for the anchor element.                                                                                                                                                               | `string`                                                                     | `undefined`                            |
-| `icon`         | `icon`          | Name of the icon to use. The `filled` version will be automatically used (if found) when the tab is `selected`.                                                                                            | `string`                                                                     | `undefined`                            |
+| `htmlTitle`    | `html-title`    | Html `title` attribute for the button.                                                                                                                                                                     | `string`                                                                     | `undefined`                            |
+| `icon`         | `icon`          | <span style="color:red">**[DEPRECATED]**</span> Use a native `<button>` instead.<br/><br/>Name of the icon to use. The `filled` version will be automatically used (if found) when the tab is `selected`.  | `string`                                                                     | `undefined`                            |
 | `label`        | `label`         | Label to show in the tab.                                                                                                                                                                                  | `string`                                                                     | `undefined`                            |
 | `orientation`  | `orientation`   | Tab orientation. Do not set this manually: `z-navigation-tabs` will handle this.                                                                                                                           | `NavigationTabsOrientation.HORIZONTAL \| NavigationTabsOrientation.VERTICAL` | `NavigationTabsOrientation.HORIZONTAL` |
 | `selected`     | `selected`      | Whether the tab is selected.                                                                                                                                                                               | `boolean`                                                                    | `false`                                |
 | `size`         | `size`          | Tab size. Do not set this manually: `z-navigation-tabs` will handle this.                                                                                                                                  | `NavigationTabsSize.BIG \| NavigationTabsSize.SMALL`                         | `NavigationTabsSize.BIG`               |
 | `tabId`        | `tab-id`        | <span style="color:red">**[DEPRECATED]**</span> Use native `id` attribute instead<br/><br/>`id` attribute of the tab. Set this id to the `aria-labelledby` attribute of the controlled `tabpanel` element. | `string`                                                                     | `undefined`                            |
-| `target`       | `target`        | Html `target` attribute for the anchor element.                                                                                                                                                            | `string`                                                                     | `undefined`                            |
 
 
 ## Events
@@ -51,13 +49,13 @@
 
 ### Depends on
 
-- [z-icon](../../../icons/z-icon)
+- [z-icon](../../components/icons/z-icon)
 
 ### Graph
 ```mermaid
 graph TD;
-  z-navigation-tab-link --> z-icon
-  style z-navigation-tab-link fill:#f9f,stroke:#333,stroke-width:4px
+  z-navigation-tab --> z-icon
+  style z-navigation-tab fill:#f9f,stroke:#333,stroke-width:4px
 ```
 
 ----------------------------------------------
